Added DevProjects to Exercises under HTML, CSS, and JavaScript, Java, PHP, Python, and Ruby. Also added it to Android and Mac and iOS.
DevProjects is a community with free real-world projects in multiple coding languages and tech stacks. The website also includes solutions from other developers and discussion board for all users to participate in. Apart from the projects, solutions, and discussion board, developers/learners can also reach out to mentors on Codementor for paid help. However, the latter bit is not compulsory or forced.